Fix `composer diagnose` warnings
Fix warnings and errors from `composer diagnose`. I chose to use
a custom type of "managed-library-collection" to indicate that this is
not any normal type of Composer package. The addition of name and
license corrected warnings. Preferred-install being in the root and not
in the config section was an error.

Change-Id: I6e412053cfc4d4a83e0c23e7e64c4662f0287a4c