Add library for use in Wikibase
This is a tiny library with a MessageReporter interface and some
simple implementations. We have been using a copy of it for many
years in Wikibase and now want to remove the copied code. We are
creating some new component which uses the library added here,
so if we don't remove the copy we'll end up with two copies of
the same code in our codebase, which would be confusing.

MessageReporter was written by me, so this is not a third party thing.

You can see the copy of this code in Wikibase at
Wikibase/lib/includes/Reporting

The library itself is at https://github.com/onoi/message-reporter

Change-Id: Id26fe8d620f375d460a9ab70819f4b194d0f6b7a