Add CirrusSearch dev dependency to require-dev
When CI runs the wmf-quibble-vendor-mysql-hhvm-docker job the phpunit
part only pulls in require-dev dependencies from mediawiki and vendor
repositories, but not from the extensions. As a workaround add
necessary require-dev lines to vendor.git itself.

The dependency added is swaggest/json-schema. This is used to validate
that log messages being sent to EventBus match the schema that they
claim.

Change-Id: I51eca32bdfd71bc7b617f499d69662614dd1270d