Update symfony/polyfill-php80 to v1.18.0
This satisfies the version dependencies of a number of symfony/*
libraries, which require symfony/polyfill-php80 ^1.15.  This is a
follow up to I75f1c62db8e301e687e0645bcac6299d6e94d321 and we've gone
ahead and pinned the specific version of symfony/polyfill-php80 as was
done in that patch.

Change-Id: I6b2f2d939fa661bca05f14b797ad57c4a1edc988