Run `composer update --no-dev` with composer 2.2.11
It appears that the current version of `composer/` has been generated
with composer 2.1.9 (aka, it doesn't contain the phpdoc changes to
InstalledVersions.php which landed upstream in
d1c8a4d1b419889a4f53ab58173bc875a102b6f0 and was released in 2.1.10).
The README.md for this repository says we should be using "the latest
version of 2.2.x of composer".  That's 2.2.11 at this point in time,
so I've regenerated the `composer/` directory using this version of
composer.

Change-Id: If8181aa9a70c81a8f480835bd3a977ada2c12062