Initial audit of mck89/peast package
This is a dependency-free package aimed to replace the JSMinPlus
library currently forked and committed within MediaWiki core.

JSMinPlus, unlike the name might suggest, is not the minifier
currently used by ResourceLoader in MediaWiki. For that we use,
JavaScriptMinifier (part of the wikimedia/minify package).

Instead, JSMinPlus is currently used as a linter (through tokenizing
and thus validating the syntax) for MediaWiki:Common.js and gadgets
to avoiud deploying broken code to client browsers.

https://github.com/mck89/peast
https://packagist.org/packages/mck89/peast

Bug: T75714
Change-Id: I43d4619a32e37eb42e1aaa55a1f602962609c52b