Add wikimedia/json-codec 2.2.1
Used by wikimedia/parsoid in new patch set
I2983d27d40d7c5951f17d62a8f3dd8b6ac6abdef.

This library began as a librarization of the code in JsonCodec code
in mediawiki-core, but has diverged for the needs of the Parsoid
project.  The interfaces it defines can be made compatible with
the JsonCodec in core (T346829).

Bug: T346829
Bug: T214994
Change-Id: I86a846d013a6342ee833d1b1c945c8795f578b55