Bump wmde/php-vuejs-templating to 2.2.0-beta.2
More functionality that we want to start using in Wikibase (follows up
change I9d43a495fb, commit c52a90c102). There’s probably still more to
come, and we might discover that the interface is not quite what we
want, so for now this is a beta version; I’m assuming that, at least for
“our own” (Wikimedia) libraries, using non-stable versions is acceptable
(compare wikimedia/parsoid, which regularly uses alpha versions, and
league/oauth2-server, where we’ve had to use v9.0.0-alpha for years).

Bug: T396445
Change-Id: I5f2fd83f73b78d8c91c7831af9bead4f7635eb5e