Add wikimedia/codex v0.7.1 (Codex PHP)
wikimedia/codex depends on krinkle/intuition, but it doesn't use it when
running in MediaWiki. We don't want to install krinkle/intuition here,
so ignore it.

Exclude the i18n directory, since it contains files that are only used
by Intuition when running outside of MW, but are not used when running
in MW.

Bug: T379662
Change-Id: Idc9ea14f8bb2eb2f225f4d2f8be184dcb85f47ce